cryptohome: Optional enabling of NVRAM TPM 2.0 live test
Add a command line flag to the cryptohome-tpm-live-test tool
which allows to enable the NVRAM test on TPM 2.0.

It's been disabled by default as this test requires the TPM 2.0
owner password to be known to the system (e.g., tpm_manager),
and as there's no simple way for the test tool to inject the
password even if it knew that.

With this CL, the test can be manually enabled. This will be
particularly useful for the platform_CryptohomeTpmLiveTest
autotest, as we can be sure that the owner password must stay
there in the test workflow.

BUG=chromium:832226
TEST=check that the NvramTest test gets executed and succeeds by
     running this on a TPM 2.0 Chromebook after the TPM is cleared
     and has just been re-owned:
     cryptohome-tpm-live-test --tpm2_use_system_owner_password

Change-Id: Iee5043dd0fe4dfa841e174ae0acb5a4e58bd22c9
Reviewed-on: https://chromium-review.googlesource.com/1073456
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>