authpolicy: Cache kdc_ip, dc_name and workgroup
Authpolicy is fairly slow since it makes a lot of Samba calls, many of
them are almost always unnecessary. For instance, the Kerberos key
distribution center (KDC) IP, the domain controller (DC) name or the
workgroup are pretty much constants. This CL caches them in memory.

A little complication arises because the KDC IP is fetched in the same
query as the server time, which always changes. Yet, the server time is
only needed for machine password handling, so only enforce up-to-date
server time for that.

If the values ever change, calls to Active Directory might start
failing. The Kerberos ticket might not refresh and policy might fail to
fetch. If that happens, logging out and back in will fix it since
authpolicyd is then restarted.

BUG=chromium:845829
TEST=Enter 'authpolicy_debug 3' in crosh to enable debug logs.
     Log in, reload policy a few times. Look at /var/log/authpolicy.log.
     The following should only show up twice (for user and for device):
     Executing /usr/bin/net 'ads' 'lookup'
     Executing /usr/bin/net 'ads' 'info'
     Executing /usr/bin/net 'ads' 'workgroup'

Change-Id: I7fe572cc2e8ba63a8dd6026292d657f11e5d46a2
Reviewed-on: https://chromium-review.googlesource.com/1073249
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>