vm_tools: vsh: default to empty user for container vsh
On the client side, set the default value of the user flag to be empty,
which indicates that vshd should run as the current user in a container.
Set user to "chronos" when connecting to a VM.

Also add a check for this in vshd, since this will let us remove the hack
in the client code later.

BUG=chromium:819324
TEST=manual

Change-Id: I77ba3bd9a607d8440f59b4ecedb401d0a8055cd3
Reviewed-on: https://chromium-review.googlesource.com/1074088
Commit-Ready: Stephen Barber <smbarber@chromium.org>
Tested-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>