vm_tools: sommelier: add support for NV12 shm format
This adds support for NV12 as a wl_shm format when shm driver
is is 'virtwl-dmabuf' and host compositor supports this dmabuf
format.

Suppot for this format is possible as we copy into dmabufs when
using the 'virtwl-dmabuf' driver. The mmap utilities have been
updated to support multi-plane formats. The code is still limited
to 2 planes and the same bpp for all planes as that's all we need
for NV12. It can be made even more flexible if needed in the
future.

NV12 is useful for efficient video playback. It allows a software
decode pipeline to output NV12 and a single copy before planes
can be scanned out using an HW overlay.

BUG=chromium:837209
TEST=gst-launch-1.0 playbin video-sink=waylandsink uri=http://distribution.bbb3d.renderfarming.net/video/mp4/bbb_sunflower_1080p_60fps_normal.mp4

Change-Id: I6f4a4cb48ef9912c46d082606cda87c6bd18e0b6
Reviewed-on: https://chromium-review.googlesource.com/1074552
Commit-Ready: David Reveman <reveman@chromium.org>
Tested-by: David Reveman <reveman@chromium.org>
Reviewed-by: David Reveman <reveman@chromium.org>