init: Fix the check if clobber has to run after clear tpm
The previous version was checking if the TPM is owned using the cryptohome
command. That doesn't work when the chrome OS starts, and it was always
returning false. Now it's fixed by checking the contents of device/owned
file.

Also the location of install attributes file that was checked in previous
version was wrong, since /home/.shadow/ is now mounted yet at that point.
This is fixed now too.

BUG=chromium:701292
TEST=Build a image that boots in verified mode and check clobber state logs
after a normal boot and after a boot in devmode switched to verified mode.

Change-Id: I4d3d593fa29e42ea1cd94682962c7e07ea10b695
Reviewed-on: https://chromium-review.googlesource.com/1075274
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>