cryptohome: Challenge credentials - decryption operation
Implementation of decryption of challenge credentials. It will
be used both for the new vault keyset creation (as the freshly
created secrets are encrypted and need to be immediately decrypted)
and for the existing vault keyset decryption.

The implementation consists of combining the signature of
salt with the secret which is unsealed (decrypted via a
challenge-response dance involving the TPM).

BUG=chromium:842791
TEST=none (unit tests will be in a follow-up)

Change-Id: Iacfc8704a89a8dfec8803eab22bf05ee127911e0
Reviewed-on: https://chromium-review.googlesource.com/1074748
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>