cryptohome: LE Credential sync - Part 1
This is the first of a two-part change to add support for Log
Replay functionality for Low Entropy (LE) credentials when the on-disk
state and LE Backend state have gone out of sync.

This CL adds support to verify the hash cache state, and possibly
rebuild it if it's detected to be stale. This functionality is added
in the Sync() function of LECredentialManager(). It is expected that
Sync() will be called before every LECredentialManager public operation.

BUG=chromium:809710, chromium:794010
TEST=1. cros_run_unit_tests
     2. Run following operations on eve (with a PinWeaver Cr50)
        a. Create an LE Credential.
        b. Check LE credential.
        c. Corrupt the leafcache using hexedit and then reboot.
        d. Make sure we can perform further LE operations, including
           checking the same LE credential, and removing it.

Change-Id: Iab5844dfb10f3e880d6b4fbd1160af245d206963
Reviewed-on: https://chromium-review.googlesource.com/1056400
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Prashant Malani <pmalani@google.com>