vm_tools: Refactor part of concierge into cicerone
This refactors all of the container communication that is done in
concierge into cicerone. It is still expected that Chrome will only be
calling into concierge, so all cicerone calls get proxied from concierge
into cicerone. After this CL lands, then we can update Chrome to using
cicerone instead of concierge where relevant. After Chrome OS is then
uprev'd to that version of Chrome, we can then remove all the proxying
to cicerone from concierge.

This CL also needs crrev.com/c/1069772 to land and Chrome OS uprev'd to
include it before it is merged so that Chrome receives the cicerone
ContainerStarted signal and understands it.

BUG=chromium:042845
TEST=Verified terminal startup, open url, icons and apps work on eve

Change-Id: I5172431eb3ae800975b9084014e803f9531910fd
Reviewed-on: https://chromium-review.googlesource.com/1080148
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>