shill: Add cmdline flag for spawning VPN clients in minijail
This flag can be used by code in vpn/ to decide whether to spawn VPN
client programs in a minijail, which will be necessary if we are running
the shill process tree as non-root. Doing this through a command line
flag enables a simple change in the shill init job to back out of
sandboxing shill if something goes wrong.

BUG=chromium:649417
TEST=tested as part of larger sandbox shill debug CL

Change-Id: I0955bde21909e62354487aaf1f7229db02405b51
Reviewed-on: https://chromium-review.googlesource.com/1081354
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>