shill: cellular: respect authentication method in APN property
A user-specific APN is provided to shill through setting the APN
property of a cellular service object. The APN property is a dictionary
that contains the APN name and other optional entries such as username,
password, and authentication method. CL:857742 and CL:857743 added the
support for APN authentication method specified in the mobile operator
database.  This CL updates shill to respect the authentication method in
the user-specified APN property. Once the corresponding support is added
to the Chrome network settings UI, a user will be able to specify the
authentication method for a custom APN.

BUG=b:68680371
TEST=Run unit tests.
TEST=Verify that shill preserves the authentication method specified
     through the APN property of a cellular service.

Change-Id: Idd1eee22007a93ef0a597cfed231753805d17aa8
Reviewed-on: https://chromium-review.googlesource.com/1081764
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>