arc: Exit ip_helper normally when read fails with errno ECONNRESET
In CL:1065120, ip_helper was modified to send a NO-OP message upon setup to
it's parent process to confirm that the socket connection is still alive.
However, when the parent process exits later on (and closes the socket), the
child, which normally would get a 0 upon calling read() and exit normally,
now gets -1 with errno ECONNRESET. This happens when one side of a socket
connection writes, the other side exits without reading, and the first side
attempts to read.

This change has ip_helper exit normally when the read fails with ECONNRESET.

BUG=chromium:845967
TEST=Put a log statement in the if clause. Log statement was reached and
     arc-networkd stopped crashing.

Change-Id: I44b201761e29645abf23a10a08a630b079fc341d
Reviewed-on: https://chromium-review.googlesource.com/1086116
Commit-Ready: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Abhishek Bhardwaj <abhishekbh@chromium.org>