shill: give shill access to debugfs files and /proc/net/ip_conntrack
This CL is in preparation to run the shill process tree as a non-root
user.

Adjust perms on miscellaneous files that need to be accessed by shill as part of
performing network configuration and monitoring. Shill is able to access
networking-related files in /sys/class/net and /proc/sys/net through the
use of runtime capabilities, so these 3 exceptions should be the only
ones we need to make for sandboxing shill.

BUG=chromium:649417, chromium:804941
TEST=tested as part of larger sandbox shill debug CL

Change-Id: I3b1ab8e10cc7fa87fbfba03272c5921ab29f8947
Reviewed-on: https://chromium-review.googlesource.com/1086211
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>