libchromeos-ui: do not disable GPU sandbox on VMs
The --gpu-sandbox-failures-fatal=no is passed to the Chrome process when
it is detected that we are not running on actual Chrome OS hardware.
This is to work around the fact that the LLVMpipe Mesa driver, used on
Betty VMs, creates threads before sandboxing and thus fails the sandbox
sanity tests.

However this behavior also triggered on actual hardware that does not
use Coreboot as its firmware, which cause quite some confusion on e.g.
b/79891439.

The LLVM thread creation has been delayed to post-sandoxing in
crrev.com/c/1078212, thus we do not need this workaround anymore. This
patch removes it.

BUG=b:79891439
TEST=Verified that Cheza booted without issue
CQ-DEPEND=CL:1078212

Change-Id: I47846d238f54a1e9749f239b0f215368f3dbbb93
Reviewed-on: https://chromium-review.googlesource.com/1086818
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>