authpolicy: Remove unused field
Removes last_auth_error from ActiveDirectoryUserStatus as it's not
neeeded anymore. It was supposed to detect whether the Active Directory
server can be accessed, but that turns out to be flaky since the
network might be switching during AuthenticateUser(). Calling
PingServer() in GetUserStatus() is more robust.

Also, Chrome never used it, so it's safe to remove.

CQ-DEPEND=CL:1090711

BUG=chromium:849318
TEST=trybots

Change-Id: I16d49d795becf813dab88357b3f611c3459cfbe4
Reviewed-on: https://chromium-review.googlesource.com/1086795
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>