authpolicy: Fix password change flow
In case a user logs in from the PODs screen (so-called "offline logon",
but actually being online), and their password is expired,
AuthenticateUser() returns ERROR_PASSWORD_EXPIRED, but it's not picked
up by Chrome since Chrome doesn't wait for the result on the PODs
screen. Chrome signs the user in and calls GetUserStatus(). If the TGT
is not valid, the GetUserStatus() call just returns success, but does
not set the password status, so Chrome never gets notified of this.

This CL checks the last auth error and sets the password status to
- PASSWORD_EXPIRED if last auth error was ERROR_PASSWORD_EXPIRED and
- PASSWORD_CHANGED if last auth error was ERROR_BAD_PASSWORD.
The rationale for the latter is that the password must have been an old,
valid password since the Cryptohome mount succeeded.

Also adds a debug flag for logging the GetUserStatus() result.

BUG=chromium:849318
TEST=1) Set account password to expired in AD, log on from PODs screen.
     You should get a notification that the password expired and relog
     should take you to online logon, where you can change the password.
     2) Change password in AD, log on from PODs screen. You should get a
     notification that the password changed on the server and relog
     should take you to online logon, where you have to enter the old
     and new password.

Change-Id: Iabc6fe0822e59f741c1cd365231f0f8215df3b02
Reviewed-on: https://chromium-review.googlesource.com/1090711
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>