shill: start running shill process tree in a minijail
Kick off shill sandboxing (we'll see if it sticks). This simple change to
the init script as well as the autotest change to
security_SandboxedServices are the only 2 things that need to be
reverted to go back to running the shill process tree as root.

CQ-DEPEND=CL:1082958
BUG=chromium:649417
TEST=observed shill process tree running as non-root. no autotests
break.

Change-Id: Ice5ed2cbbc64281ba8f136e5157023ff355245d6
Reviewed-on: https://chromium-review.googlesource.com/1087540
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>