cryptohome: Reduce use of SecureBlob in BootAttributes.
Polymorphic usage of SecureBlob is problematic (see crbug.com/728047).
BootAttributes is using SecureBlobs in places that don't really require
it. In preparation for making SecureBlob not inherit from brillo::Blob
(which would break some of these usages), reduce use of SecureBlob.

BootAttributes are meant to be tamper-evident, not read-protected.
Therefore we don't really need to read the data into a SecureBlob.
Public keys also don't need to be SecureBlobs, nor does any signed
data. Replace all of these to be plain Blobs.

Signatures technically don't need to be SecureBlobs either (they're
just encrypted hashes), but the existing TPM API makes it annoying
to pass Blobs as signatures. To avoid having to do extra conversions,
keep signatures as SecureBlobs.

The Service class exposes methods that used to take a SecureBlob (and
now take Blobs), but those methods are not used outside of the class,
so they can also be updated in this CL.

BUG=chromium:728047
TEST=cryptohome unit tests still pass on samus and eve.
TEST=eve boots.
TEST=kevin boots.

Change-Id: Ia1d111f8562d38a3fa226ef4832a7e4fd5d4aa59
Reviewed-on: https://chromium-review.googlesource.com/1088991
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Xiaoyong Zhou <xzhou@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>