vm_tools: maitred: allow users to access dmesg
We don't have a reason to lock out dmesg access to users, so lets drop
this.  The general LXD production docs recommend locking it down, but
that's for a general container environment where there might be a lot
of unrelated containers running.  In our setup, we have isolated VM's
whose only job is to run containers and let users do whatever.  There
shouldn't be anything private in the kernel log to hide (for now).

BUG=None
TEST=precq passes

Change-Id: I88982a6b0edbe1b371b263c8bf1fd5c7c7e7e4bf
Reviewed-on: https://chromium-review.googlesource.com/1087847
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jingkui Wang <jkwang@google.com>
Reviewed-by: Dylan Reid <dgreid@chromium.org>