cryptohome: Fix one missing polymorphic use of SecureBlob.
CL:1037363 missed this case in a rebase.

While the TPM key doesn't really need a SecureBlob, it's read from disk
and then immediately passed to TPM::LoadWrappedKey, which does take the
key as a SecureBlob. We could avoid the ReadFileToSecureBlob call if we
used ReadFile and then converted the Blob into a SecureBlob, but it's
cleaner to change the read function.

BUG=chromium:728047
TEST=cryptohome unit tests still pass on samus and eve.
TEST=eve boots.
TEST=kevin boots.

Change-Id: Iae2e50ac6b26a6e1bd80e4c4c8e03de970d9d4d7
Reviewed-on: https://chromium-review.googlesource.com/1090813
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>