cryptohome: Add UMA reporting for LE Check Credentials
This CL adds UMA reporting for the result of Low Entropy
CheckCredential() calls. It also adds the relevant enums required
for logging errors, and constructing the various histogram strings.

Logging is added for the InsertCredential() and Sync() operations as
a proof of concept. Based on feedback, the other operations will have
logging added to them in a follow up CL.

A follow up CL will add corresponding histograms for the remaining
LE operations.

BUG=chromium:809749
TEST=cros_run_unit_tests

Change-Id: I6e5259ba4d357b87a6f2e657aed7947684cebb80
Reviewed-on: https://chromium-review.googlesource.com/1093957
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>