cecservice: delay source address setting until a message is sent
Fix a minor problem where we might fail to send a message in transient
state after TV gets connected (the physical address is configured) and
before logical address is set. In this state we would always use
'UNREGISTERED' logical address as a source, which is not correct.
The change is to delay setting of a source address on a message
until we actually send it.

BUG=chromium:828820
TEST=Run unit test and also sanity checked on a actual daemon using
 dbus-send command.

Change-Id: I9d2ddfd5c5424eafa8ca453cc52112d0c8dbc908
Reviewed-on: https://chromium-review.googlesource.com/1097124
Commit-Ready: Dariusz Marcinkiewicz <darekm@google.com>
Tested-by: Dariusz Marcinkiewicz <darekm@google.com>
Reviewed-by: Felix Ekblom <felixe@chromium.org>