vpn: Reland "Set RPC identifiers from command line" (CL:1094615)
Original change's description:
> vpn: Set RPC identifiers from command line
>
> This CL is in preparation to run the shill process tree as a non-root
> user.
>
> Currently the shill binary spawns l2tpipsec_vpn.cc with empty
> environment variables except for 2 necessary ones that act as RPC
> connection/object identifiers for shill to communicate with
> l2tpipsec_vpn. Since we will soon be spawning l2tpipsec_vpn in a
> minijail (which doesn't support setting environment vars for spawned
> process) need to pass these value on the command line instead.
> Additionally, since previous versions of the shill code cleared the
> environment variables before adding the 2 needed ones, we now do that
> from within the l2tpipsec_vpn binary.
>
> CQ-DEPEND=CL:1039882
> BUG=chromium:649417
> TEST=tested as part of larger sandbox shill debug CL

Slight modification is that we skip trying to set the env vars from the
command line if they have already been set by the parent process. This
way, shill doesn't have to spawn l2tpipsec_vpn with special arguments
when spawning through ExternalTask::Start. On the other hand,
ExternalTask::StartInMinijail will still pass the variables on the
command line and won't set up the env vars correctly, so in that case
clear the vars from the parent and set the 2 meaningful ones from the
command line. A good reason to have l2tpipsec_vpn work in both of these
cases is so we can switch back and forth between spawning it in a
minijail and not, in the case that we need to revert enabling sandboxing
the shill process tree.

BUG=chromium:649417
TEST=tested as part of larger sandbox shill debug CL

Change-Id: Ie874ecadde7aee77772ccc9822a94fde89012fbd
Reviewed-on: https://chromium-review.googlesource.com/1099491
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>