arc-setup: Use kRootUid and Gid for the host-side dalvik-cache dir
This allows us to use upstream Android's app_process/app_main.cpp
(zygote startup code) as-is in P and master. The directory with the
new permissions is exposed to the container only while the stateless
mini container is running, and unmounted before the container starts
executing non-verified code, so no 3rd party process will be writing
to the directory. Also, even in a very unlikely case where a process
in the mini container writes random files to the directory for any
reasons, the written content is completely deleted (with host root
privilege) when 'arc-setup --setup' is executed to start the mini
container next time resetting the host-side directory to the known
safe state.

BUG=b:111849512
TEST=ARC starts even with ag/4351571
TEST=Add a file to the directory, confirm it's deleted before mini
 container starts next time.
TEST=security_StatefulPermissions
CQ-DEPEND=CL:1150868

Change-Id: I9d6bb7e0c0b150111da8ded4aee27b1f554a428f
Reviewed-on: https://chromium-review.googlesource.com/1102055
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>