vm_tools: Support launching apps with file arguments
This CL adds support for the |files| argument in
LaunchContainerApplicationRequest, which will be passed through to the
container and used to launch applications with the given file arguments.
For now, applications which only support launching with a single file
will only use the first file in the list, instead of being launched
multiple times.

BUG=chromium:822513
TEST=Manually verified this works in combination with
https://crrev.com/c/1100606 and https://crrev.com/c/1103995.
CQ-DEPEND=CL:1103995
Change-Id: I87d7bab6aad8cebe6e14ed29596f1cb413ad08e6
Reviewed-on: https://chromium-review.googlesource.com/1104060
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Timothy Loh <timloh@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>