chromeos-common-script: Fix install_hybrid_mbr
install_hybrid_mbr was not working as sfdisk was not in the path.
Using sudo adds /sbin before running the command.
Also, if the argument is a loop device, sudo is mandatory.

The problem was hidden as overlays were re-defining install_hybrid_mbr()
with proper sudo calls.

Newer version of sfdisk (2.32) would wipe other partition tables by
default when another tabe is modified. Add option -w never to prevent
wipe when adding PMBR.

BUG=chromium:853582
TEST=Check beaglebone work with

Change-Id: I80b9bf8aa4ef80490ec0d70cda2b2782977f498b
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1104699
Reviewed-by: Mike Frysinger <vapier@chromium.org>