imageloader: add an Manifest class
This is a refactor change:
1. add Manifest class
2. move manifest parsing code from Component to Manifest class.

Reason to refactor:
Currently imageloader supports component only which has a unified code
path to parse manifest file and do operations. This is not an ideal
situation when we are adding support for DLC which skips signature
verification (for imageloader.json)&table hash verifications. So we
extract common code path
for parsing manifest file to a Manifest class that can be used by both
Components&DLC.

BUG=chromium:850155
TEST=unittest, platform_AddPrinter.epson

Change-Id: Icbf29da6146674529fc88c6dc21c4bf1d746f447
Reviewed-on: https://chromium-review.googlesource.com/1106807
Commit-Ready: Xiaochu Liu <xiaochu@chromium.org>
Tested-by: Xiaochu Liu <xiaochu@chromium.org>
Reviewed-by: Xiaochu Liu <xiaochu@chromium.org>