chromeos-config: Correct impl of single-element string list
At present a string list with a single element is treated as a single
string in the device-tree implementation. This is not correct. It should
be a list with a single element.

Correct this and add a test to cover this case for both FDT and yaml
impls. The test uses the GetFirmwareInfo() API since this is the only one
that currently uses string lists.

BUG=chromium:853801
TEST=FEATURES=test sudo -E emerge chromeos-config-tools

Change-Id: Icb255fbddc64eb763daf3aaa8f32e4679fd0ca57
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1107844
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: YH Lin <yueherngl@chromium.org>