cryptohome: Reduce needless SecureBlob usage in attestation code.
Polymorphic usage of SecureBlob is problematic (see crbug.com/728047).
In preparation for making SecureBlob not inherit from Blob (which will
break some of these usages), reduce unnecessary use of SecureBlob for
non-secrets.

There is one use of SecureBlob for the modulus of a public key, and
another one for the input to a decryption function. These don't need
to be SecureBlobs.

Also fix some style issues.

BUG=chromium:728047
TEST=cryptohome unit tests still pass.

Change-Id: I8166a078954ba46cd33706a17e27a794f1701005
Reviewed-on: https://chromium-review.googlesource.com/1112475
Commit-Ready: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Yves Arrouye <drcrash@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>