vm_tools: concierge: Set hostname in SSH keys response
This sets the hostname to be the <container>-<vm>-local in the SSH keys
response. This is so that we can safely transition to
<container>.<vm>.test soon without introducing temporary breakage. Once
Chrome has updated to this change and Chrome OS is uprev'd to that
version of Chrome the we can do the change to .test.

BUG=chromium:825010
TEST=Builds
CQ-DEPEND=CL:1113908

Change-Id: Ieb61885bcb160c0b01d7b297bf7c3f994682b19a
Reviewed-on: https://chromium-review.googlesource.com/1113913
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>