cryptohome: Modify Platform::Mount to use mount flags.
Currently, Platform::Mount uses a base set of mount flags.
Instead of creating a new function for mounting with flags,
we modify the signature of Mount since there are not many
direct users of the mount function. This is in preparation
of using Mount() with flags in mount-encrypted.

- Rename kDefaultMountOptions to kDefaultMountFlags to
  disambiguate between the arguments for flags and options.

BUG=chromium:808303
TEST=compiles

Change-Id: I69af6dbf765dfa311f45b8e98909dec446450d83
Signed-off-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1116339
Commit-Ready: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>