shill: wifi: don't crash; just retry wpa_supplicant CreateInterface
We should not crash simply because wpa_supplicant failed to bring up the
interface. The supplicant can fail for a number reasons, including (most
commonly) a Wifi driver failing to reach IFF_UP. This is often permanent
(e.g., driver bug or hardware failure), and we definitely shouldn't put
shill into an abort/respawn loop, as this just prevents stable
connectivity via other means (e.g., ethernet or cellular).

Instead, let's just schedule an arbitrary interval at which shill will
retry bringing up the interface, and try that a few times.

If the device eventually disappears (e.g., Wifi firmware crashes and
kernel device completely unregisters itself), the entire WiFi device
will get removed and invalidate any outsanding callbacks.

If wpa_supplicant restarts, we'll see OnSupplicantVanish() which will
terminate the retry loop.

If the user disables WiFi in the tray, we'll terminate via
'Device::enabled() == false'.

If we don't succeed within this arbitrary retry loop, we
non-persistently disable the interface. This allows the user to
re-enable the device, if somehow the device recovers later.

I also add a metric (Network.Shill.WiFi.SupplicantAttempts) to log
either success-after-N-attempts or a large ("Max") value for failure.
This can help us continue to identify device populations that are having
lots of driver failures in the field, as well as evaluate whether this
is truly transient (e.g., succeeds after a few retries) or permanent.

NB: after studying more of the existing device issues that hit this
LOG(FATAL) behavior in shill, I suspect that these failures do *not*
tend to be transient and instead require some other intervention to
recover the device. (We're working on these interventions elsewhere.) If
the metrics tell us the retries are never useful, perhaps we can scale
this back to simply stop trying after the first failure.

BUG=chromium:698457
TEST=force a mac80211 Wifi driver to fail its start() method a few times
     -- see shill retry to bring it up
TEST=manually re-enable Wifi (e.g., in UI) and see shill try to
     reconnect a failed device
TEST=unit tests
TEST=watch chrome://histograms

Change-Id: I1c735b474609dfe0de437f6eb65b1b74fae33aab
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1383253
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>