shill: RTNL: fixup bounds checking in decoding
The fuzzing framework noticed two heap overruns in RTNL parsing.

For NLMSG_ERROR, we validate that the header fits in our buffer, but we
don't validate that the 'nlmsgerr' does. Fix that.

For DeocdeNdUserOption(), our bounds check isn't big enough; beyond the
first element, we're trusting that the headers contain correct length
values, and we only bounds check against them. Extend the initial bounds
checking to account for all the bits we're expecting to be packed into
this message.

BUG=chromium:856866
TEST=build for x86 with USE="asan fuzzer", then run
     /usr/libexec/fuzzers/rtnl_handler_fuzzer
TEST=network_Ipv6SimpleNegotiation

Change-Id: I0bb289208f08d971355be10033389877fc600716
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1116254
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>