init: turn off ext4 quota for builds before 10756
Build 10756.0.0 enabled quota on /mnt/stateful_partition in order
to support ARC++. Unfortunately, in the test lab, we try to preserve
stateful on a host when installing the next test image. If that next
image is older than 10756.0.0:
1) the mount command will fail when the next image boots
2) the failure is reported as "corrupted stateful"
3) machine will then powerwash to recover
4) the provision step fails because python is not present on stateful
5) the machine will get "repaired" successfully

This has already happened to a pool of machines that was already low
on devices. The "temporary" loss of additional machines prevented the
corresponding paladin from running test suites.

Turning off quota has to be done before we boot the buid image
which knows nothing about quota. And we only need/want to do
this for test lab machines. And we can only run tune2fs AFTER
the stateful partition is unmounted.

BUG=chromium:854278
TEST=manual

Change-Id: Id3e86c482857cb67710441b697855e35b8404173
Reviewed-on: https://chromium-review.googlesource.com/1116286
Commit-Ready: Grant Grundler <grundler@chromium.org>
Tested-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@google.com>
Reviewed-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>