Add Android's Reserved GID to resgid
Android added GID 1065 to resgid:
https://android.googlesource.com/platform/system/core/+/1fefb9f1294d3e270f1711f41e7c77a61213742b/libcutils/include/private/android_filesystem_config.h#128

We want to add that GID to ChromeOS' ext4 resgid, as elaborated in
http://go/arc++:reserved-space. With current mapping,
adding Android's GID 1065 to ChromeOS's ext4 resgid means adding GID
1065 + shift (655360). However, ext4 resgid is unfortunately still a
16-bit integer (this is based on observation. The code is probably: https://chromium.googlesource.com/chromiumos/third_party/kernel/+/20b3692ddbf17eda938a899c387d941a2743edfc/fs/ext4/ext4.h#1170). Hence, adding 1065 + 655360 is interpreted as adding GID
1065 (because of modulo 65536). Hence, we need to map the 1065 to a
special GID that is not used in ChromeOS yet that is
less than 2^16=65536. In this case, we are assigning 1065 to 20119,
which is a GID in 20100-29999 range that can be allocated for this
purpose.

BUG=b:111529716
TEST=Restarted device and GID=20119 is in ext4 resgid (through
dumpe2fs). And also tested that after Android reserved gid is added in resgid (which
is container's GID=1065), exhausting disk space by a non-system Android
app should not crash the container.
CQ-DEPEND=CL:1145151
Change-Id: I2b61f5213f6872e50d8d09fd40b605f9a6a7446a
Reviewed-on: https://chromium-review.googlesource.com/1143175
Commit-Ready: Risan <risan@chromium.org>
Tested-by: Risan <risan@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>