cryptohome: Refactor error handling in Mount::DecryptVaultKeyset
Preparatory CL for removal of duplicated logic of vault keyset loading
from this method.

The goal is to separate out the error handling logic (i.e., which kind
of MountError should be used when the keyset loading or decryption
fails). After doing that, the next step would be the removal of the
duplicate low-level loading mechanism (in favor of the already existing
means provided by VaultKeyset+HomeDirs).

No behavior changes are expected. Only some error messages for error
scenarios (when the loading/decryption fails) became changed and/or
more granular.

BUG=chromium:832395
TEST=existing tests,
     manual test: log in under a new account, wait until the TPM is
     initialized, log out, log in again, log out, try to log in again
     with a wrong password - and verify that the user cryptohome was
     retained

Change-Id: I6b77fb1a3e6d7ba562c3ab8a457cbed28608f683
Reviewed-on: https://chromium-review.googlesource.com/1118553
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>