hermes: Rename std::vector<uint8_t> to DataBlob
The data exchange between the eSIM and the SM-DP+ server is encrypted by
either end using keys that are unknown to the Lpd. DataBlob conveys that
the Lpd API shouldn't care about what data is being transferred between
the two endpoints except for the header of the packets.

BUG=b:80591752
BUG=chromium:847619
TEST=emerge-cheza hermes

Change-Id: I4f123175fe8929087ce4de38996662ea646433a0
Reviewed-on: https://chromium-review.googlesource.com/1119502
Commit-Ready: Jacob Rutherford <jruthe@google.com>
Tested-by: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Ben Chan <benchan@chromium.org>