vm_tools: maitred: Add Mount9P RPC
Add a new RPC for mounting 9P file systems.  We cannot reuse the
existing mount RPC for this because we want to connect to the server
over the vsock transport, which the kernel doesn't directly support via
mount flags.

Instead maitre'd connects to the server itself and mounts the file
system with the trans=fd option, using the fd of the connected socket.

BUG=chromium:703939
TEST=Use maitred_client to mount a 9p server

Change-Id: I333d872ff527ac664739972d163c5f0b1fdb9540
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1121579
Reviewed-by: Stephen Barber <smbarber@chromium.org>